pyproject.toml
[project]
name = "todoapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.3",
    "fastapi[standard]>=0.128.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.46",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
    "ty>=0.0.15",
]
